M5.5 chunk 3: full FFI conversion; docs.
Every expressible export and mirrored type now goes through the generator: logging (buck_log in macro form; new Rust-side public LogLevel enum as the C# mirror's source of truth, compile-time-pinned to the log crate's discriminants), the whole platform surface (pump, bulk event poll, window create/destroy/set-title/size with typed Rid handles, the keycode probes), and PlatformEventRaw (its hand layout-seam trio replaced by the generated pair). The emitter grew string-in, Span<T> out-buffer, and named-tuple multi-return wrapper emission alongside these first consumers; usize returns surface as checked C# int. Hand LogLevel.cs/PlatformEventRaw.cs and the converted imports are deleted; zero "keep in sync" comments remain in the tree; the hand-written FFI residue is exactly the fn-pointer family + buck_last_error_message + the C# callback infrastructure.

Recorded normalizations, both sides regenerated in lockstep: events_poll capacity u32 -> usize; keycode_name no longer reports needed length on a too-small-buffer error; exception context strings take the grep-exact symbol name.

Docs: ARCHITECTURE's FFI section rewritten around binding generation (including the pinned error philosophy); PLAN gains the M5.5 milestone section and the deferred entry's as-built annotations.

+++ src/core/tests/window_live.rs
@@ -38,7 +38,7 @@ fn create_pump_destroy_round_trip() {
         let poll = unsafe {
             buckminster_core::platform::buck_platform_events_poll(
                 buf.as_mut_ptr(),
-                buf.len() as u32,
+                buf.len(),
                 &mut written,
                 &mut remaining,
             )
